# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "API for the Ludus decision operating system.",
    "title": "Ludus API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/workspaces/{workspaceId}/simulations/{graphId}/versions": {
      "get": {
        "description": "Paged graph-version history, newest first (canonical 10-api row).",
        "operationId": "list_graph_versions_api_workspaces__workspaceId__simulations__graphId__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "graphId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Graphid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "workspaceId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspaceid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Graph Versions Api Workspaces  Workspaceid  Simulations  Graphid  Versions Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Graph Versions",
        "tags": [
          "workspaces",
          "simulations"
        ]
      }
    },
    "/api/workspaces/{workspaceId}/simulations/{graphId}/versions/{graphVersionId}": {
      "get": {
        "description": "Exact immutable graph version with nodes and edges (canonical row).",
        "operationId": "get_graph_version_api_workspaces__workspaceId__simulations__graphId__versions__graphVersionId__get",
        "parameters": [
          {
            "in": "path",
            "name": "graphId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Graphid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "graphVersionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Graphversionid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "workspaceId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspaceid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Graph Version Api Workspaces  Workspaceid  Simulations  Graphid  Versions  Graphversionid  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Graph Version",
        "tags": [
          "workspaces",
          "simulations"
        ]
      }
    },
...
  },
...
}